sync: MAJOR upstream 144.0 → 146.0.2

+++ js/src/vm/NativeObject.cpp
@@ -6,7 +6,6 @@
 
 #include "vm/NativeObject-inl.h"
 
-#include "mozilla/Casting.h"
 #include "mozilla/CheckedInt.h"
 #include "mozilla/Maybe.h"
 
@@ -2356,7 +2355,7 @@ bool js::NativeGetElement(JSContext* cx, Handle<NativeObject*> obj,
   RootedId id(cx);
 
   if (MOZ_LIKELY(index >= 0)) {
-    if (!IndexToId(cx, index, &id)) {
+    if (!IndexToId(cx, uint32_t(index), &id)) {
       return false;
     }
   } else {
